Move function call outside protocol, into the viewmodel
