+++ WooCommerce/Classes/ViewRelated/Orders/Order Details/Address Edit/EditOrderAddressForm.swift
@@ -2,6 +2,7 @@ import Combine
 import SwiftUI
 import UIKit
 import Yosemite
+import Experiments
 
 /// Hosting controller that wraps an `EditOrderAddressForm`.
 ///
@@ -89,6 +90,8 @@ struct EditOrderAddressForm<ViewModel: AddressFormViewModelProtocol>: View {
 
     @Environment(\.safeAreaInsets) var safeAreaInsets: EdgeInsets
 
+    let isSearchCustomersEnabled = DefaultFeatureFlagService().isFeatureFlagEnabled(.orderCreationSearchCustomers)
+
     var body: some View {
         Group {
             ScrollView {
@@ -147,11 +150,13 @@ struct EditOrderAddressForm<ViewModel: AddressFormViewModelProtocol>: View {
                 })
             }
             ToolbarItem(placement: .automatic) {
-                Button(action: {
-                    viewModel.callToCustomerAction()
-                }, label: {
-                    Image(systemName: "magnifyingglass")
-                })
+                if isSearchCustomersEnabled {
+                    Button(action: {
+                        viewModel.callToCustomerAction()
+                    }, label: {
+                        Image(systemName: "magnifyingglass")
+                    })
+                }
             }
             ToolbarItem(placement: .confirmationAction) {
                 navigationBarTrailingItem()
